+++ src/templates/install.ps1
@@ -33,8 +33,11 @@ $websiteLocation = $config.WebsiteLocation
 $url = $config.Url
 $blueprintFolderPath = $config.BlueprintFolderPath
 $backupDir = $config.BackupDir
-$isDelivery = $config.IsDelivery.ToLower() -eq "true"
 $itemsDirectory = $config.ItemsDirectory
+$configFolder = "$scriptPath\configs"
+$role = $config.Role
+$environment = $config.Environment
+$isDelivery = $config.IsDelivery.ToLower() -eq "true"
 
 if(-not $websiteLocation) {
     Write-Error "Website location is not set!"
@@ -75,23 +78,20 @@ else {
 
 
 Write-Output "Copy content of $scriptPath\website to  $websiteLocation"
-cp  "$scriptPath\website\*" "$websiteLocation\" -Recurse
+cp "$scriptPath\website\*" "$websiteLocation\" -Recurse
 
+Install-LoftyEnvironment -WebRoot $websiteLocation `
+    -Environment $environment `
+    -Role $role `
+    -BlueprintFolderPath $blueprintFolderPath `
+    -ConfigFolder $configFolder
 
-Write-Output "Restore unmanaged files from $blueprintFolderPath to $websiteLocation"
-
-if($blueprintFolderPath) {
-    
-    cp  "$blueprintFolderPath\*" "$websiteLocation\" -Recurse -Force
-    
-}
 
 Write-Output "Starting IIS app pool $appPoolName"
 Start-WebAppPool $appPoolName
 
 if (-not $isDelivery)
 {
-    $configFolder = "$scriptPath\configs"
     Install-AppItems $url "$itemsDirectory\items" "$scriptPath\tempAppItems" $configFolder $websiteLocation -AppItemsZipPath "$scriptPath\app.zip" 
     Install-DefaultItems $url "$itemsDirectory\defaultItems" "$scriptPath\tempDefaultItems" -DefaultItemsZipPath "$scriptPath\appDefault.zip"
 }